Move main to seperate file
