+++ parser.py
@@ -693,7 +693,7 @@ def simple_stringify(query: Query) -> str:
 
 # Example usage in the main block
 if __name__ == "__main__":
-    input_query = '([word="A"]) ; (([word="B" word="X"] | [word="C"])) ; (([word="D"] | [word="E"] | [word="F"]))'
+    input_query = '[word="A"] ; ([word="B"] | [word="C"]) ; ([word="D"] | [word="E"] | [word="F"]) ; ([word="X"] | [word="Y"] | [word="Z"] | [word="W"])'
     
     if len(sys.argv) > 1:
         input_query = sys.argv[1]
