+++ parser/disjunction.py
@@ -26,7 +26,7 @@ class DisjunctionQuery(Query):
     def __eq__(self, other: Query) -> bool:
         if not isinstance(other, DisjunctionQuery):
             return False
-        return self.queries == other.queries
+        return set(self.queries) == set(other.queries)
     
     def __hash__(self) -> int:
         return hash(tuple(self.queries))
